copilotkit-pydantic/first-party-mcp-servers/confluence/requirements.txt
# Confluence MCP Server Requirements

# MCP Framework
fastmcp>=0.3.0

# Confluence SDK
atlassian-python-api>=3.41.0

# Caching
cachetools>=5.3.0

# Web Server
uvicorn[standard]>=0.27.0

# Type Safety
pydantic>=2.0.0

